rework of main charge into a class, start of ethernet interface

+++ lib/constants/CCUParams.h
@@ -2,7 +2,7 @@
 #define __CCUPARAMS_H__
 namespace CCUParams{
   const double target_voltage = 3.1; //per cell
-  const int max_AC_current = 15; //need to add the DC current setpoint probably 30 amps
+  const float max_AC_current = 15; //need to add the DC current setpoint probably 30 amps
   const float ideal_cell_temp = 40;
 };
 #endif
